refactor(examples): consolidate upgrade pipeline command
